move functions to srt/configs/update_config.py and srt/model_loader/weight_utils.py
