+++ python/sglang/srt/layers/moe/fused_moe_triton/layer.py
@@ -18,13 +18,13 @@ from sglang.srt.layers.quantization.base_config import (
     QuantizationConfig,
     QuantizeMethodBase,
 )
+from sglang.srt.model_loader.weight_utils import narrow_padded_param_and_loaded_weight
 from sglang.srt.utils import (
     _process_weight_after_loading,
     cpu_has_amx_support,
     get_bool_env_var,
     is_cpu,
     is_hip,
-    narrow_padded_param_and_loaded_weight,
     set_weight_attrs,
 )
