+++ python/sglang/srt/utils.py
@@ -52,7 +52,6 @@ from json import JSONDecodeError
 from multiprocessing.reduction import ForkingPickler
 from pathlib import Path
 from typing import (
-    TYPE_CHECKING,
     Any,
     Callable,
     Dict,
@@ -89,13 +88,6 @@ from triton.runtime.cache import (
     default_dump_dir,
     default_override_dir,
 )
-
-if TYPE_CHECKING:
-    from sglang.srt.configs.load_config import LoadConfig
-    from sglang.srt.configs.model_config import ModelConfig
-
-
-DEFAULT_MOE_PADDING_SIZE = 32
 
 logger = logging.getLogger(__name__)
 
@@ -2555,165 +2547,3 @@ def align(x: int, y: int) -> int:
 # COPIED FROM DeepGEMM
 def ceil_div(x: int, y: int) -> int:
     return (x + y - 1) // y
-
-
-def may_get_weight_block_size(model_config, load_config):
-    from sglang.srt.model_loader.loader import _get_quantization_config
-    from sglang.srt.model_loader.utils import get_model_architecture
-
-    model_class, _ = get_model_architecture(model_config)
-    packed_modules_mapping = getattr(model_class, "packed_modules_mapping", {})
-
-    quant_config = _get_quantization_config(
-        model_config, load_config, packed_modules_mapping
-    )
-
-    if quant_config is not None and hasattr(quant_config, "weight_block_size"):
-        return getattr(quant_config, "weight_block_size")
-    return None
-
-
-def get_num_heads_padding_size(tp_size, weight_block_size):
-    pad_size = (
-        tp_size * 2 if tp_size % 2 == 1 and weight_block_size is not None else tp_size
-    )
-    return pad_size
-
-
-def get_moe_padding_size(weight_block_size):
-    if weight_block_size is not None:
-        # See NOTE(HandH1998): To ensure proper alignment of the block-wise quantization scales, the output_size of the weights for both the gate and up layers must be divisible by block_n.
-        assert (
-            len(weight_block_size) == 2
-        ), "Only len(weight_block_size) == 2 is supported"
-        assert (
-            weight_block_size[0] == weight_block_size[1]
-        ), "Only weight_block_size[0] == weight_block_size[1] is supported"
-
-        return weight_block_size[0]
-
-    return DEFAULT_MOE_PADDING_SIZE
-
-
-def update_intermediate_size(model_config, attr_name, intermediate_padding_size):
-    if hasattr(model_config.hf_config, attr_name):
-        attr_value = getattr(model_config.hf_config, attr_name)
-        if attr_value % intermediate_padding_size != 0:
-            from sglang.srt.layers.vocab_parallel_embedding import pad_vocab_size
-
-            attr_value = pad_vocab_size(attr_value, intermediate_padding_size)
-            setattr(model_config.hf_config, attr_name, attr_value)
-            setattr(model_config.hf_text_config, attr_name, attr_value)
-    return model_config
-
-
-def update_config(
-    model_config: ModelConfig, load_config: LoadConfig, tp_size: int
-) -> ModelConfig:
-    # Support the case where the num_attention_heads is not divisible by the TP size.
-    weight_block_size = may_get_weight_block_size(model_config, load_config)
-
-    if (
-        model_config.num_attention_heads % tp_size != 0
-        or model_config.get_total_num_kv_heads() % tp_size != 0
-    ):
-        # Compute the head_dim using the model_config.num_attention_heads before padding
-        model_config.hf_config.original_head_dim = (
-            model_config.hidden_size // model_config.num_attention_heads
-        )
-
-        model_config.hf_config.original_num_attention_heads = (
-            model_config.num_attention_heads
-        )
-        model_config.hf_text_config.original_num_attention_heads = (
-            model_config.num_attention_heads
-        )
-
-        model_config.hf_config.original_total_num_kv_heads = (
-            model_config.get_total_num_kv_heads()
-        )
-        model_config.hf_text_config.original_total_num_kv_heads = (
-            model_config.get_total_num_kv_heads()
-        )
-
-        query_heads_per_kv = (
-            model_config.num_attention_heads // model_config.get_total_num_kv_heads()
-        )
-        total_kv_heads = model_config.get_total_num_kv_heads()
-        from sglang.srt.layers.vocab_parallel_embedding import pad_vocab_size
-
-        pad_size = get_num_heads_padding_size(tp_size, weight_block_size)
-        num_key_value_heads = pad_vocab_size(total_kv_heads, pad_size)
-
-        model_config.num_key_value_heads = num_key_value_heads
-        model_config.hf_config.num_key_value_heads = num_key_value_heads
-        model_config.hf_text_config.num_key_value_heads = num_key_value_heads
-
-        num_attention_heads = num_key_value_heads * query_heads_per_kv
-        model_config.num_attention_heads = num_attention_heads
-        model_config.hf_config.num_attention_heads = num_attention_heads
-        model_config.hf_text_config.num_attention_heads = num_attention_heads
-
-    intermediate_padding_size = tp_size * get_moe_padding_size(weight_block_size)
-    model_config = update_intermediate_size(
-        model_config, "moe_intermediate_size", intermediate_padding_size
-    )
-    model_config = update_intermediate_size(
-        model_config, "intermediate_size", intermediate_padding_size
-    )
-
-    return model_config
-
-
-def get_actual_shard_size(shard_size, weight_start, weight_end):
-    if weight_end < weight_start:
-        return 0
-
-    return min(shard_size, weight_end - weight_start)
-
-
-def reset_param_data_if_needed(param_data, dim, start, length):
-    if length == 0:
-        return
-
-    assert length > 0, f"Length should be positive, but got {length}"
-
-    param_data.narrow(dim, start, length).zero_()
-    return
-
-
-def narrow_padded_param_and_loaded_weight(
-    param_data,
-    loaded_weight,
-    param_data_start,
-    weight_start,
-    dim,
-    shard_size,
-    narrow_weight=True,
-):
-    actual_shard_size = get_actual_shard_size(
-        shard_size, weight_start, loaded_weight.size(dim)
-    )
-
-    if narrow_weight:
-        if actual_shard_size > 0:
-            loaded_weight = loaded_weight.narrow(dim, weight_start, actual_shard_size)
-        else:
-            # No real data to load; create a dummy tensor filled with zeros
-            loaded_weight = torch.zeros_like(
-                param_data.narrow(dim, param_data_start, actual_shard_size)
-            )
-
-    # [Note] Reset padded weights to zero.
-    # If the actual shard size is less than the shard size, we need to reset
-    # the padded param_data to zero and then copy the loaded_weight into it.
-    reset_param_data_if_needed(
-        param_data,
-        dim,
-        param_data_start + actual_shard_size,
-        shard_size - actual_shard_size,
-    )
-
-    param_data = param_data.narrow(dim, param_data_start, actual_shard_size)
-
-    return param_data, loaded_weight
